char-array-example-ascii

+++ proj/tic-tac-toe/main.cpp
@@ -12,8 +12,8 @@ int main()
       {0, 0, 0},
       {0, 0, 0},
       {0, 0, 0} };
-  unsigned int total_rows = sizeof(board)/sizeof(board[0]);
-  unsigned int total_cols = sizeof(board[0])/sizeof(board[0][0]);
+  unsigned short total_rows = sizeof(board)/sizeof(board[0]);
+  unsigned short total_cols = sizeof(board[0])/sizeof(board[0][0]);
   
   for(unsigned int row = 0; row < total_rows;row++) {
       for(unsigned int col = 0; col < total_cols;col++) {
